Add stickyfill polyfill to fix nav behavior in IE
Resolves department-of-veterans-affairs/vets-contrib#263
