+++ config/paths.js
@@ -57,4 +57,5 @@ module.exports = {
   appTsLint: resolveApp('tslint.json'),
   publicUrl: getPublicUrl(resolveApp('package.json')),
   servedPath: getServedPath(resolveApp('package.json')),
+  ansiStyles: resolveApp('node_modules/ansi-styles'),
 };
