+++ config/webpack.config.dev.js
@@ -145,7 +145,16 @@ module.exports = {
               compact: true,
             },
           },
-
+          {
+            test: /\.(js|jsx|mjs)$/,
+            loader: require.resolve('babel-loader'),
+            options: {
+              babelrc: false,
+              compact: false,
+              presets: [require.resolve('babel-preset-react-app/dependencies')],
+              cacheDirectory: true,
+            },
+          },
           // Compile .tsx?
           {
             test: /\.(ts|tsx)$/,
