+++ config/webpack.config.prod.js
@@ -279,6 +279,16 @@ module.exports = (envName) => {
                 compact: true,
               },
             },
+            {
+              test: /\.(js|jsx|mjs)$/,
+              loader: require.resolve('babel-loader'),
+              options: {
+                babelrc: false,
+                compact: false,
+                presets: [require.resolve('babel-preset-react-app/dependencies')],
+                cacheDirectory: true,
+              },
+            },
             // Compile .tsx?
             {
               test: /\.(ts|tsx)$/,
