+++ package.json
@@ -23,7 +23,7 @@
     "babel-jest": "23.6.0",
     "babel-loader": "8.0.4",
     "babel-plugin-named-asset-import": "^0.2.3",
-    "babel-preset-react-app": "^6.1.0",
+    "babel-preset-react-app": "^7.0.0",
     "case-sensitive-paths-webpack-plugin": "2.1.1",
     "chalk": "1.1.3",
     "css-loader": "1.0.0",
@@ -71,6 +71,7 @@
     "scroll-into-view-if-needed": "^2.2.20",
     "source-map-loader": "^0.2.1",
     "stickyfill": "^1.1.1-strict",
+    "stickyfilljs": "^2.1.0",
     "style-loader": "0.23.0",
     "swagger-ui": "^3.20.4",
     "swagger-ui-themes": "^3.0.0",
