+++ src/components/index.ts
@@ -3,6 +3,7 @@ export * from './NavBar';
 export * from './Footer';
 export * from './PageContent';
 export * from './PageHero';
+export * from './ApiPage';
 import SwaggerDocs from './SwaggerDocs';
 
 export { SwaggerDocs };
