+++ src/containers/Explore.scss
@@ -18,14 +18,13 @@
     }
 
     .sticky {
-
+      top: 20px;
       display: none;
 
       @include media($medium-screen) {
 
-        display:block;
+        display: block;
         position: sticky;
-        top: 20px;
       }
     }
